Frontend: Extract UI Components
Start grouping GUI responsibilities into new package forge.frontend...

First step: forge.frontend.components.widgets
(Most of forge.gui.interfaces - only those with only frontend
responsibilities)
)
