+++ forge-gui-desktop/src/main/java/forge/screens/home/quest/VSubmenuDuels.java
@@ -8,11 +8,11 @@ import javax.swing.JPanel;
 import javax.swing.ScrollPaneConstants;
 import javax.swing.SwingConstants;
 
+import forge.frontend.components.widgets.IButton;
 import forge.gamemodes.quest.IVQuestStats;
 import forge.gui.framework.DragCell;
 import forge.gui.framework.DragTab;
 import forge.gui.framework.EDocID;
-import forge.gui.interfaces.IButton;
 import forge.localinstance.skin.FSkinProp;
 import forge.screens.home.EMenuGroup;
 import forge.screens.home.IVSubmenu;
