+++ forge-gui-desktop/src/main/java/forge/toolbox/FComboBoxWrapper.java
@@ -13,7 +13,7 @@ import javax.swing.ListCellRenderer;
 
 import com.google.common.collect.ObjectArrays;
 
-import forge.gui.interfaces.IComboBox;
+import forge.frontend.components.widgets.IComboBox;
 import forge.toolbox.FComboBox.TextAlignment;
 import forge.toolbox.FSkin.SkinFont;
