+++ forge-gui-mobile/src/forge/itemmanager/filters/AdvancedSearchFilter.java
@@ -7,7 +7,7 @@ import com.google.common.collect.Iterables;
 import forge.Forge;
 import forge.assets.FSkinImage;
 import forge.assets.TextRenderer;
-import forge.gui.interfaces.IButton;
+import forge.frontend.components.widgets.IButton;
 import forge.item.InventoryItem;
 import forge.itemmanager.AdvancedSearch;
 import forge.itemmanager.AdvancedSearch.IFilterControl;
