+++ forge-gui-mobile/src/forge/menu/FMenuItem.java
@@ -9,10 +9,10 @@ import forge.Graphics;
 import forge.assets.FImage;
 import forge.assets.FSkinColor;
 import forge.assets.FSkinColor.Colors;
+import forge.frontend.components.widgets.IButton;
 import forge.assets.FSkinFont;
 import forge.assets.TextRenderer;
 import forge.gui.UiCommand;
-import forge.gui.interfaces.IButton;
 import forge.localinstance.skin.FSkinProp;
 import forge.screens.FScreen.Header;
 import forge.toolbox.FDisplayObject;
