+++ forge-gui-mobile/src/forge/screens/planarconquest/ConquestStatsScreen.java
@@ -6,10 +6,10 @@ import forge.Forge;
 import forge.assets.FImage;
 import forge.assets.FSkinFont;
 import forge.assets.FSkinImage;
+import forge.frontend.components.widgets.IButton;
 import forge.gamemodes.planarconquest.ConquestData;
 import forge.gamemodes.planarconquest.ConquestPlane;
 import forge.gamemodes.planarconquest.IVConquestStats;
-import forge.gui.interfaces.IButton;
 import forge.model.FModel;
 import forge.screens.FScreen;
 import forge.toolbox.FComboBox;
