+++ forge-gui-mobile/src/forge/screens/quest/QuestMenu.java
@@ -9,14 +9,14 @@ import forge.deck.Deck;
 import forge.deck.DeckGroup;
 import forge.deck.FDeckEditor.DeckController;
 import forge.deck.FDeckEditor.EditorType;
+import forge.frontend.components.widgets.IButton;
+import forge.frontend.components.widgets.ICheckBox;
+import forge.frontend.components.widgets.IComboBox;
 import forge.gamemodes.quest.IVQuestStats;
 import forge.gamemodes.quest.QuestUtil;
 import forge.gamemodes.quest.data.QuestPreferences.QPref;
 import forge.gamemodes.quest.io.QuestDataIO;
 import forge.gui.FThreads;
-import forge.gui.interfaces.IButton;
-import forge.gui.interfaces.ICheckBox;
-import forge.gui.interfaces.IComboBox;
 import forge.localinstance.properties.ForgeConstants;
 import forge.menu.FMenuItem;
 import forge.menu.FPopupMenu;
