+++ forge-gui-mobile/src/forge/screens/quest/QuestStatsScreen.java
@@ -6,12 +6,12 @@ import forge.Forge;
 import forge.assets.FImage;
 import forge.assets.FSkinFont;
 import forge.assets.FSkinImage;
+import forge.frontend.components.widgets.IButton;
+import forge.frontend.components.widgets.ICheckBox;
+import forge.frontend.components.widgets.IComboBox;
 import forge.gamemodes.quest.QuestUtil;
 import forge.gamemodes.quest.bazaar.QuestItemType;
 import forge.gamemodes.quest.bazaar.QuestPetController;
-import forge.gui.interfaces.IButton;
-import forge.gui.interfaces.ICheckBox;
-import forge.gui.interfaces.IComboBox;
 import forge.model.FModel;
 import forge.screens.FScreen;
 import forge.toolbox.FCheckBox;
