+++ forge-gui-mobile/src/forge/toolbox/FButton.java
@@ -10,10 +10,10 @@ import forge.Forge;
 import forge.Graphics;
 import forge.assets.FSkinColor;
 import forge.assets.FSkinColor.Colors;
+import forge.frontend.components.widgets.IButton;
 import forge.assets.FSkinFont;
 import forge.assets.FSkinImage;
 import forge.gui.UiCommand;
-import forge.gui.interfaces.IButton;
 import forge.localinstance.skin.FSkinProp;
 import forge.toolbox.FEvent.FEventHandler;
 import forge.toolbox.FEvent.FEventType;
