+++ forge-gui-mobile/src/forge/toolbox/FLabel.java
@@ -9,11 +9,11 @@ import forge.Graphics;
 import forge.assets.FImage;
 import forge.assets.FSkinColor;
 import forge.assets.FSkinColor.Colors;
+import forge.frontend.components.widgets.IButton;
 import forge.assets.FSkinFont;
 import forge.assets.FSkinImage;
 import forge.assets.TextRenderer;
 import forge.gui.UiCommand;
-import forge.gui.interfaces.IButton;
 import forge.localinstance.skin.FSkinProp;
 import forge.toolbox.FEvent.FEventHandler;
 import forge.toolbox.FEvent.FEventType;
