+++ forge-gui-mobile/src/forge/toolbox/FTextField.java
@@ -8,8 +8,8 @@ import forge.Forge.KeyInputAdapter;
 import forge.Graphics;
 import forge.assets.FSkinColor;
 import forge.assets.FSkinColor.Colors;
+import forge.frontend.components.widgets.ITextField;
 import forge.assets.FSkinFont;
-import forge.gui.interfaces.ITextField;
 import forge.menu.FMenuItem;
 import forge.menu.FPopupMenu;
 import forge.toolbox.FEvent.FEventHandler;
