+++ forge-gui/src/main/java/forge/frontend/components/widgets/ICheckBox.java
@@ -1,4 +1,4 @@
-package forge.gui.interfaces;
+package forge.frontend.components.widgets;
 
 public interface ICheckBox extends IComponent {
     boolean isSelected();
