+++ forge-gui/src/main/java/forge/frontend/components/widgets/IProgressBar.java
@@ -1,4 +1,4 @@
-package forge.gui.interfaces;
+package forge.frontend.components.widgets;
 
 public interface IProgressBar {
     void setDescription(String s0);
