+++ forge-gui/src/main/java/forge/frontend/components/widgets/ITextComponent.java
@@ -1,4 +1,4 @@
-package forge.gui.interfaces;
+package forge.frontend.components.widgets;
 
 public interface ITextComponent extends IComponent {
     String getText();
