+++ forge-gui/src/main/java/forge/frontend/components/widgets/ITextField.java
@@ -1,4 +1,4 @@
-package forge.gui.interfaces;
+package forge.frontend.components.widgets;
 
 public interface ITextField extends ITextComponent {
     boolean requestFocusInWindow();
