+++ forge-gui/src/main/java/forge/gamemodes/planarconquest/ConquestChaosBattle.java
@@ -6,6 +6,7 @@ import java.util.Set;
 
 import forge.LobbyPlayer;
 import forge.deck.Deck;
+import forge.frontend.components.widgets.IButton;
 import forge.game.GameType;
 import forge.game.GameView;
 import forge.gamemodes.planarconquest.ConquestPreferences.CQPref;
@@ -13,7 +14,6 @@ import forge.gamemodes.quest.QuestEventDifficulty;
 import forge.gamemodes.quest.QuestEventDuel;
 import forge.gamemodes.quest.QuestEventDuelManager;
 import forge.gamemodes.quest.QuestWorld;
-import forge.gui.interfaces.IButton;
 import forge.gui.interfaces.IGuiGame;
 import forge.gui.interfaces.IWinLoseView;
 import forge.item.PaperCard;
