+++ forge-gui/src/main/java/forge/gamemodes/planarconquest/ConquestController.java
@@ -28,6 +28,7 @@ import forge.deck.CardPool;
 import forge.deck.Deck;
 import forge.deck.DeckSection;
 import forge.deck.io.DeckStorage;
+import forge.frontend.components.widgets.IButton;
 import forge.game.GameRules;
 import forge.game.GameType;
 import forge.game.GameView;
@@ -37,7 +38,6 @@ import forge.gamemodes.planarconquest.ConquestPreferences.CQPref;
 import forge.gamemodes.quest.BoosterUtils;
 import forge.gui.FThreads;
 import forge.gui.GuiBase;
-import forge.gui.interfaces.IButton;
 import forge.gui.interfaces.IGuiGame;
 import forge.gui.interfaces.IWinLoseView;
 import forge.item.PaperCard;
