+++ forge-gui/src/main/java/forge/gamemodes/quest/QuestUtil.java
@@ -22,6 +22,7 @@ import forge.LobbyPlayer;
 import forge.card.CardEdition;
 import forge.card.CardRules;
 import forge.deck.Deck;
+import forge.frontend.components.widgets.IButton;
 import forge.game.GameRules;
 import forge.game.GameType;
 import forge.game.player.RegisteredPlayer;
@@ -35,7 +36,6 @@ import forge.gamemodes.quest.data.QuestAssets;
 import forge.gamemodes.quest.data.QuestPreferences;
 import forge.gui.FThreads;
 import forge.gui.GuiBase;
-import forge.gui.interfaces.IButton;
 import forge.gui.interfaces.IGuiGame;
 import forge.gui.util.SGuiChoose;
 import forge.gui.util.SOptionPane;
