+++ forge-gui/src/main/java/forge/itemmanager/AdvancedSearch.java
@@ -29,6 +29,7 @@ import forge.deck.Deck;
 import forge.deck.CardPool;
 import forge.deck.DeckProxy;
 import forge.deck.DeckSection;
+import forge.frontend.components.widgets.IButton;
 import forge.game.GameFormat;
 import forge.game.keyword.Keyword;
 import forge.gamemodes.planarconquest.ConquestCommander;
@@ -39,7 +40,6 @@ import forge.gamemodes.quest.QuestWorld;
 import forge.gui.FThreads;
 import forge.gui.GuiBase;
 import forge.gui.UiCommand;
-import forge.gui.interfaces.IButton;
 import forge.gui.util.SGuiChoose;
 import forge.gui.util.SOptionPane;
 import forge.item.InventoryItem;
