+++ forge-gui/src/main/java/forge/itemmanager/SFilterUtil.java
@@ -16,9 +16,9 @@ import forge.card.CardRulesPredicates;
 import forge.card.ColorSet;
 import forge.card.MagicColor;
 import forge.deck.DeckProxy;
+import forge.frontend.components.widgets.IButton;
 import forge.game.GameFormat;
 import forge.gamemodes.quest.data.StarRating;
-import forge.gui.interfaces.IButton;
 import forge.item.InventoryItem;
 import forge.item.PaperCard;
 import forge.itemmanager.SItemManagerUtil.StatTypes;
