+++ forge-gui/src/main/java/forge/itemmanager/SItemManagerUtil.java
@@ -10,8 +10,8 @@ import com.google.common.base.Predicate;
 import forge.card.CardRules;
 import forge.card.CardRulesPredicates;
 import forge.deck.DeckProxy;
+import forge.frontend.components.widgets.IComboBox;
 import forge.gui.GuiBase;
-import forge.gui.interfaces.IComboBox;
 import forge.item.InventoryItem;
 import forge.localinstance.skin.FSkinProp;
 import forge.localinstance.skin.IHasSkinProp;
