+++ forge-gui/src/main/java/forge/model/FModel.java
@@ -32,6 +32,7 @@ import forge.deck.CardArchetypeLDAGenerator;
 import forge.deck.CardRelationMatrixGenerator;
 import forge.deck.io.DeckPreferences;
 import forge.download.AutoUpdater;
+import forge.frontend.components.widgets.IProgressBar;
 import forge.game.GameFormat;
 import forge.game.GameType;
 import forge.game.card.CardUtil;
@@ -50,7 +51,6 @@ import forge.gamemodes.tournament.TournamentData;
 import forge.gui.FThreads;
 import forge.gui.GuiBase;
 import forge.gui.card.CardPreferences;
-import forge.gui.interfaces.IProgressBar;
 import forge.item.PaperCard;
 import forge.itemmanager.ItemManagerConfig;
 import forge.localinstance.achievements.*;
